Add buttons to Grant/Revoke user privileges
