+++ indico/htdocs/js/indico/jquery/declarative.js
@@ -75,7 +75,6 @@
             var params = $this.data('params') || {};
             var paramsSelector = $this.data('params-selector');
             var update = $this.data('update');
-            var ajaxRequest = $this.data('ajax-request') !== undefined;
             var dialog = $this.data('ajax-dialog') !== undefined;
             var reload = $this.data('reload-after');
             if (!$.isPlainObject(params)) {
@@ -151,7 +150,7 @@
                 }
 
                 // Handle html update
-                if (update || ajaxRequest) {
+                if (update) {
                     $.ajax({
                         method: method,
                         url: url,
