+++ indico/modules/events/management/blueprint.py
@@ -18,9 +18,7 @@ from __future__ import unicode_literals
 
 from indico.modules.events import event_management_object_url_prefixes
 from indico.modules.events.management.controllers import (RHDeleteEvent, RHLockEvent, RHUnlockEvent,
-                                                          RHShowNonInheriting, RHEventProtection,
-                                                          RHGrantSubmissionRights, RHGrantModificationRights,
-                                                          RHRevokeSubmissionRights)
+                                                          RHShowNonInheriting, RHEventProtection)
 from indico.web.flask.wrappers import IndicoBlueprint
 
 
@@ -32,11 +30,6 @@ _bp.add_url_rule('/delete', 'delete', RHDeleteEvent, methods=('GET', 'POST'))
 _bp.add_url_rule('/lock', 'lock', RHLockEvent, methods=('GET', 'POST'))
 _bp.add_url_rule('/unlock', 'unlock', RHUnlockEvent, methods=('POST',))
 _bp.add_url_rule('/protection', 'protection', RHEventProtection, methods=('GET', 'POST'))
-_bp.add_url_rule('/protection/grant-submission', 'grant_submission_rights', RHGrantSubmissionRights, methods=('POST',))
-_bp.add_url_rule('/protection/grant-modification', 'grant_modification_rights', RHGrantModificationRights,
-                 methods=('POST',))
-_bp.add_url_rule('/protection/revoke-submission', 'revoke_submission_rights', RHRevokeSubmissionRights,
-                 methods=('POST',))
 
 for object_type, prefixes in event_management_object_url_prefixes.iteritems():
     if object_type == 'subcontribution':
