+++ indico/modules/events/persons/blueprint.py
@@ -16,7 +16,8 @@
 
 from __future__ import unicode_literals
 
-from indico.modules.events.persons.controllers import RHPersonsList, RHEmailEventPersons
+from indico.modules.events.persons.controllers import (RHPersonsList, RHEmailEventPersons, RHGrantSubmissionRights,
+                                                       RHGrantModificationRights, RHRevokeSubmissionRights)
 from indico.web.flask.wrappers import IndicoBlueprint
 
 _bp = IndicoBlueprint('persons', __name__, template_folder='templates', virtual_template_folder='events/persons',
@@ -24,3 +25,8 @@ _bp = IndicoBlueprint('persons', __name__, template_folder='templates', virtual_
 
 _bp.add_url_rule('/persons/', 'person_list', RHPersonsList)
 _bp.add_url_rule('/persons/email', 'email_event_persons', RHEmailEventPersons, methods=('POST',))
+_bp.add_url_rule('/persons/grant-submission', 'grant_submission_rights', RHGrantSubmissionRights, methods=('POST',))
+_bp.add_url_rule('/persons/grant-modification', 'grant_modification_rights', RHGrantModificationRights,
+                 methods=('POST',))
+_bp.add_url_rule('/persons/revoke-submission', 'revoke_submission_rights', RHRevokeSubmissionRights,
+                 methods=('POST',))
