+++ indico/modules/events/persons/templates/management/person_list.html
@@ -22,7 +22,6 @@
 {% endblock %}
 
 {% block persons %}
-
     {% if num_no_account %}
         {% call message_box('warning') %}
             {% trans count = num_no_account -%}
